+++ src/Qir/riqir/sim/Cargo.toml
@@ -5,7 +5,7 @@ authors = ["Microsoft"]
 edition = "2021"
 
 [dependencies]
-qir-runtime = { path = "../runtime" }
+qir-stdlib = { path = "../stdlib" }
 rand = "0.8.5"
 rustc-hash = "1.1.0"
 num-complex = "0.4"
