+++ src/Qir/riqir/stdlib/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "qir-runtime"
+name = "qir-stdlib"
 version = "0.1.0"
 authors = ["Microsoft"]
 edition = "2021"
